+++ internal/worker/unpin.go
@@ -39,9 +39,9 @@ func (w *UnpinWorker) Start(ctx context.Context) error {
 func (w *UnpinWorker) handleMessage(ctx context.Context, body []byte) error {
 	cid := string(body)
 
-	// 更新状态为 Unpinning
+	// Mark as Unpinning
 	err := w.store.Update(ctx, cid, func(r *store.PinRecord) error {
-		r.Status = int32(store.StatusUnpinning)
+		r.Status = store.StatusUnpinning
 		r.UnpinStartAt = time.Now().UnixMilli()
 		return nil
 	})
@@ -50,14 +50,14 @@ func (w *UnpinWorker) handleMessage(ctx context.Context, body []byte) error {
 		return err
 	}
 
-	// 执行 unpin
+	// Execute unpin
 	if err := w.ipfs.PinRm(ctx, cid); err != nil {
 		if err.Error() == "pin/rm: not pinned or pinned indirectly" {
-			// CID 未被 pin，记录日志并返回成功
+			// Not pinned already; treat as success
 			log.Printf("CID %s is already unpinned", cid)
-			// 更新状态为 UnpinSucceeded
+			// Mark as UnpinSucceeded
 			err = w.store.Update(ctx, cid, func(r *store.PinRecord) error {
-				r.Status = int32(store.StatusUnpinSucceeded)
+				r.Status = store.StatusUnpinSucceeded
 				r.UnpinSucceededAt = time.Now().UnixMilli()
 				return nil
 			})
@@ -66,19 +66,19 @@ func (w *UnpinWorker) handleMessage(ctx context.Context, body []byte) error {
 				return err
 			}
 
-			// 删除过期索引
+			// Delete expire index entry
 			if err := w.store.DeleteExpireIndex(ctx, cid); err != nil {
 				log.Printf("Failed to delete expire index for %s: %v", cid, err)
-				// 不返回错误，因为主要操作已经成功
+				// Log only; main action succeeded
 			}
 			return nil
 		}
 		return w.handleUnpinError(ctx, cid, err)
 	}
 
-	// 更新状态为 UnpinSucceeded
+	// Mark as UnpinSucceeded
 	err = w.store.Update(ctx, cid, func(r *store.PinRecord) error {
-		r.Status = int32(store.StatusUnpinSucceeded)
+		r.Status = store.StatusUnpinSucceeded
 		r.UnpinSucceededAt = time.Now().UnixMilli()
 		return nil
 	})
@@ -87,10 +87,10 @@ func (w *UnpinWorker) handleMessage(ctx context.Context, body []byte) error {
 		return err
 	}
 
-	// 删除过期索引
+	// Delete expire index entry
 	if err := w.store.DeleteExpireIndex(ctx, cid); err != nil {
 		log.Printf("Failed to delete expire index for %s: %v", cid, err)
-		// 不返回错误，因为主要操作已经成功
+		// Log only; main action succeeded
 	}
 
 	return nil
@@ -103,12 +103,12 @@ func (w *UnpinWorker) handleUnpinError(ctx context.Context, cid string, err erro
 		r.UnpinAttemptCount++
 
 		if r.UnpinAttemptCount >= int32(w.cfg.Workers.MaxRetries) {
-			r.Status = int32(store.StatusDeadLetter)
+			r.Status = store.StatusDeadLetter
 			return nil
 		}
 
-		// 重试
-		r.Status = int32(store.StatusScheduledForUnpin)
+		// Retry
+		r.Status = store.StatusScheduledForUnpin
 		return nil
 	})
 }
